yolov5-fastapi/requirements2.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
aiofiles==0.7.0
    # via -r requirements.in
asgiref==3.4.1
    # via uvicorn
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.7
    # via requests
click==8.0.1
    # via uvicorn
cycler==0.10.0
    # via matplotlib
fastapi[standard]==0.68.1
    # via -r requirements.in
h11==0.12.0
    # via uvicorn
httptools==0.2.0
    # via uvicorn
idna==3.2
    # via requests
kiwisolver==1.3.2
    # via matplotlib
matplotlib==3.4.3
    # via
    #   -r requirements.in
    #   seaborn
# numpy==1.21.2
    # via
    #   matplotlib
    #   opencv-python-headless
    #   pandas
    #   scipy
    #   seaborn
    #   torchvision
opencv-python-headless==4.5.3.56
    # via -r requirements.in
# pandas==1.3.3
    # via
    #   -r requirements.in
    #   seaborn
pillow==8.3.2
    # via
    #   -r requirements.in
    #   matplotlib
    #   torchvision
pydantic==1.8.2
    # via fastapi
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-dotenv==0.19.0
    # via uvicorn
python-multipart==0.0.5
    # via -r requirements.in
pytz==2021.3
    # via pandas
pyyaml==5.4.1
    # via uvicorn
requests==2.26.0
    # via -r requirements.in
# scipy==1.7.1
    # via seaborn
seaborn==0.11.2
    # via -r requirements.in
six==1.16.0
    # via
    #   cycler
    #   python-dateutil
    #   python-multipart
starlette==0.14.2
    # via fastapi
# torch==1.9.1
    # via
    #   -r requirements.in
    #   torchvision
# torchvision==0.10.1
    # via -r requirements.in
tqdm==4.62.3
    # via -r requirements.in
typing-extensions==3.10.0.2
    # via
    #   pydantic
    #   torch
urllib3==1.26.7
    # via requests
uvicorn[standard]==0.15.0
    # via -r requirements.in
uvloop==0.16.0
    # via uvicorn
watchgod==0.7
    # via uvicorn
websockets==10.0
    # via uvicorn
